Tally minor update

+++ esp32_Companion_v1_Tally_1_copy_20250604231235/esp32_Companion_v1_Tally_1_copy_20250604231235.ino
@@ -1,5 +1,5 @@
 bool debug = false;
-const int version = 1;
+const float version = 1.0;
 
 //////////////////////// Wifi Manager library/////////////////////////////////////////
 #include <WiFiManager.h> // https://github.com/tzapu/WiFiManager
